changed test cases
